[refactor] DB テーブル定義をschemaに統合 (#52)
